feat(plugins): split @hawtio/react package into core and plugins
This is required to export plugins types to external plugin developers.

Splitting packages also demands exporting of util functions so that they
can be shared with the plugins package. For now, util is exported with
tsup multi entries, which means util cannot share any states with the
core, that is, only side-effect free functions can be exported util.

Due to circular dependencies between core and plugins, it's hard to move
everything in plugins into @hawtio/react-plugins. Thus, the common parts
-- shared, connect, jmx, and rbac -- are still left in @hawtio/react for
now. Ideally, we should move all of them to @hawtio/react-plugins as well
after circular dependencies are resolved.