feat(shared,jmx): support serialising long to string in JSON response from Jolokia
* Upgrade jolokia.js to 2.0.3 (that supports 'serializeLong' option)
* Create JMX preferences and provide 'Serialize long to string' option,
  which can affect Jolokia requests within JMX plugin

This fix limits the 'serializeLong=string' option to JMX, because the
change of the bahaviour might cause unexpected errors in other plugins.
JMX is a generic MBeans manipulation plugin, so it should make sense to
have such an option plugin-wide.

Each plugin should decide whether it's worth applying the option or not.

Fix #292